+++ generators/app/templates/react/components/App.jsx
@@ -1,7 +1,18 @@
 import React, { useState } from 'react';
-import DevicesPage from './DevicePage.jsx';
+<% if(isZenithBased) {%>
+<% if(isDriveAddin) {%>
+import DutyStatusLogs from './DutyStatusLogs.jsx';
+<%} else {%>
+import DevicesPage from './DevicesPage.jsx';
+<%}%>
+<%} else {%>
+import DevicePage from './DevicePage.jsx';
+<%}%>
 import GeotabContext from '../contexts/Geotab';
 import Logger from '../utils/logger';
+<% if(isZenithBased) {%>
+import '@geotab/zenith/dist/index.css'
+<%}%>
 
 const App = ({ geotabApi, geotabState, appName }) => {
   const logger = Logger(appName);
@@ -10,7 +21,15 @@ const App = ({ geotabApi, geotabState, appName }) => {
   return (
     <>
       <GeotabContext.Provider value={[context, setContext]}>
+        <% if(isZenithBased) {%>
+        <% if(isDriveAddin) {%>
+        <DutyStatusLogs />
+        <%} else {%>
         <DevicesPage />
+        <%}%>
+        <%} else {%>
+        <DevicePage />
+        <%}%>
       </GeotabContext.Provider>
     </>
   );