+++ generators/app/templates/react/typeScript/drive/Tab2Content.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+export const Tab2Content = () => {
+    return <>
+        <div className="zda-app__sections">
+            <section className="zda-section">
+                <h3 className="zda-section__header heading-02-mobile-drive">Section heading</h3>
+                <p className="zda-section__content body-02-short-mobile-drive">Tab 2</p>
+            </section>
+        </div>
+    </>;
+};